feat: store monetary values as integer cents and make gas percentage configurable - Convert monetary Float fields to Int (centavos) in Prisma schema: pixPrice, swilePrice, snapshotPixPrice, snapshotSwilePrice, totalValue, unitValue, GeneralCost.value - Migration uses ROUND(value * 100) para preservar dados existentes - Add .int() to all monetary Zod schemas in shared/schemas/ - Add Config model as singleton to hold gasPercentage - Replace hardcoded GAS_PERCENTAGE = 0.05 in purchases.service.ts with DB lookup + Math.round - Expose GET /config and PATCH /config for runtime configuration

+++ shared/schemas/financial/financialReportResponse.schema.ts
@@ -3,27 +3,27 @@ import { z } from 'zod'
 const dishBySize = z.object({
   size: z.string(),
   quantity: z.number().int(),
-  revenue: z.number(),
+  revenue: z.number().int(),
 })
 
 const dishReport = z.object({
   name: z.string(),
   quantity: z.number().int(),
-  revenue: z.number(),
+  revenue: z.number().int(),
   bySize: z.array(dishBySize),
 })
 
 const paymentMethodReport = z.object({
   quantity: z.number().int(),
-  value: z.number(),
+  value: z.number().int(),
 })
 
 export const financialReportResponse = z.object({
-  revenue: z.number(),
-  cost: z.number(),
-  profit: z.number(),
+  revenue: z.number().int(),
+  cost: z.number().int(),
+  profit: z.number().int(),
   profitMarginPercent: z.number(),
-  toReceive: z.number(),
+  toReceive: z.number().int(),
   byMethod: z.object({
     Pix: paymentMethodReport,
     Swile: paymentMethodReport,
